Allow logger to output to cout or a file.
This is done by doing some shenanigans
with the output stream. Since you can't
assign them for various reasons, it is
better to do it in the OOP style.

ostream << values_here

is defined, and both std::cout and ofstream are
either instances or derivatives of ofstream - so
we can use some pointers.