Make the log_file setting a class function
This way, multiple loggers that exist in the
program will all be writing to the same log
file, and each won't have to have their own
pointer or buffer to the same stream. (likely
not good to have the same file open from multiple
spots...)

Since the logfile is set by a class method, the
main function of the client's program can then
specify the logfile without having to instantiate
a logger.