Add the latte class to the demo code.
This part of the demo is especially nice since
it shows the order of ctor and dtor calls -
for a latte obj, first the coffee ctor is called
then the latte ctor is called to initialize the
obj. Then, during destruction, the latte dtor
is called, then the coffee dtor is called.

When an inherited method is called on the latte
obj, we also see that the function called is
the one defined in the coffee class.